Implement relative mode

+++ compute/modes_test.go
@@ -6,29 +6,29 @@ import (
 )
 
 func TestMakeModes(t *testing.T) {
-	testcases := []struct{
-		opcode int
-		want modes
+	testcases := []struct {
+		opcode int64
+		want   modes
 	}{
 		{
 			opcode: 1,
-			want: modes{},
+			want:   modes{},
 		},
 		{
 			opcode: 101,
-			want: modes{1},
+			want:   modes{1},
 		},
 		{
 			opcode: 201,
-			want: modes{2},
+			want:   modes{2},
 		},
 		{
 			opcode: 2001,
-			want: modes{0, 2},
+			want:   modes{0, 2},
 		},
 		{
 			opcode: 12001,
-			want: modes{0, 2, 1},
+			want:   modes{0, 2, 1},
 		},
 	}
 	for _, tc := range testcases {
